fix(헤더) : whiteheader -> (default)header, header -> transparentHeader
