+++ src/app/[country]/page.tsx
@@ -1,6 +1,6 @@
 import '@/styles/font.css';
 import * as styles from '@/styles/country/page.css';
-import { WhiteHeader } from '@/components/header/Header';
+import Header from '@/components/header/Header';
 import BoardPreview from '@/components/country/BoardPreview';
 import CountryBackground from '@/components/country/CountryBackground';
 import RestaurantPreview from '@/components/country/RestaurantPreview';
@@ -10,7 +10,7 @@ import SchedulePreview from '@/components/country/SchedulePreview';
 const Page = () => {
   return (
     <main>
-      <WhiteHeader />
+      <Header />
       <div className={styles.pageContainer}>
         <CountryBackground />
         <div className={styles.contentContainer}>
