+++ src/app/[country]/post/page.tsx
@@ -1,7 +1,7 @@
 import Image from 'next/image';
 import '@/styles/font.css';
 import * as country from '@/styles/country/page.css';
-import { WhiteHeader } from '@/components/header/Header';
+import Header from '@/components/header/Header';
 import CountryBackground from '@/components/country/CountryBackground';
 import PostCardList from '@/components/country/post/PostCardList';
 import CountryWrite from '@/components/common/CountryWrite';
@@ -9,14 +9,14 @@ import CountryWrite from '@/components/common/CountryWrite';
 const Page = () => {
   return (
     <main>
-      <WhiteHeader />
+      <Header />
       <div className={country.pageContainer}>
         <CountryBackground />
         <div className={country.contentContainer}>
           <PostCardList />
         </div>
       </div>
-      <CountryWrite type="post"/>
+      <CountryWrite type="post" />
     </main>
   );
 };
