+++ src/app/signup/page.tsx
@@ -1,12 +1,10 @@
 import '@/styles/font.css';
 import NonProfileHeader from '@/components/header/NonProfileHeader';
-import { pageContainer } from '@/styles/country/page.css';
 
 const Page = () => {
   return (
     <>
       <NonProfileHeader />
-      <div className={pageContainer}></div>
     </>
   );
 };
